feat(rust): rename vault -> vault_worker. add vault with convenience constructors
