+++ implementations/rust/ockam/ockam_vault_sync_core/Cargo.toml
@@ -15,6 +15,7 @@ that talk to the same Vault implementation without need for synchronization prim
 
 [features]
 default = []
+software_vault = ["ockam_vault"]
 
 [dependencies]
 ockam_core = { path = "../ockam_core", version = "0.11.0" }
