+++ implementations/rust/ockam/ockam_vault_sync_core/src/lib.rs
@@ -23,6 +23,7 @@ mod signer;
 mod symmetric_vault;
 mod vault;
 mod vault_sync;
+mod vault_worker;
 mod verifier;
 
 pub use asymmetric_vault::*;
@@ -34,4 +35,5 @@ pub use signer::*;
 pub use symmetric_vault::*;
 pub use vault::*;
 pub use vault_sync::*;
+pub use vault_worker::*;
 pub use verifier::*;
