+++ implementations/rust/ockam/ockam_vault_sync_core/src/vault_worker/response_message.rs
